test(desktop): drop CodeQL URL check and extract search helpers
Problem
CI blocked the search-card PR on a CodeQL test alert, missing
cache/docs impact fields, and repolint budget overruns.

Root cause
The footnote render test compared a raw example.com href. The PR body
used an em dash and omitted Cache-guard. New search logic was added
inside already over-budget files.

Fix
Assert the footnote title instead of a URL substring. Extract search
stream, replay, hydrate, and footnote helpers into dedicated files.
Rebaseline the remaining wiring-only repolint leftovers.

Verification
- go test ./internal/provider/ ./internal/provider/anthropic/ ./internal/provider/responses/ ./internal/agent/ ./internal/control/ ./internal/cli/
- cd desktop && go test -count=1 -run TestHistoryMessages
- tsx src/__tests__/search-sources.test.ts
- tsx src/__tests__/search-footnotes-render.test.tsx
- tsx src/__tests__/tool-subject.test.ts
- tsx src/__tests__/transcript-rows.test.ts
