+++ desktop/app.go
@@ -5742,9 +5742,7 @@ func (state *historyMessageConvertState) convertHistoryMessage(
 			hm.SubmitText = replay
 		}
 	}
-	if m.Role == provider.RoleAssistant && len(m.ServerSearch) > 0 {
-		hm.ServerSearch = historyServerSearch(m.ServerSearch)
-	}
+	hm.ServerSearch = historyServerSearch(m.ServerSearch)
 	if (m.Role == provider.RoleAssistant || m.LocalOnly) && len(m.ToolCalls) > 0 {
 		hm.ToolCalls = make([]HistoryToolCall, len(m.ToolCalls))
 		for i, tc := range m.ToolCalls {
@@ -5968,20 +5966,6 @@ func cloneHistoryMessages(in []HistoryMessage) []HistoryMessage {
 }
 
 const historyToolPreviewLimit = 2_000
-
-// historyServerSearch copies card-visible search fields and drops encrypted
-// replay payloads. Those stay on the session message for the next provider turn.
-func historyServerSearch(in []provider.ServerSearchCall) []provider.ServerSearchCall {
-	out := make([]provider.ServerSearchCall, 0, len(in))
-	for _, search := range in {
-		copied := provider.ServerSearchCall{ID: search.ID, Query: search.Query}
-		if len(search.Results) > 0 {
-			copied.Results = append([]provider.ServerSearchHit(nil), search.Results...)
-		}
-		out = append(out, copied)
-	}
-	return out
-}
 
 func historyToolCall(tc provider.ToolCall, args string, result provider.Message) HistoryToolCall {
 	call := HistoryToolCall{
