+++ desktop/frontend/src/lib/searchTranscript.ts
@@ -1,0 +1,80 @@
+import { mergeSearchSources, parseSearchSources, searchSourcesFromHistory, type SearchSource } from "./searchSources";
+import type { Item } from "./useController";
+
+type SearchState = {
+  currentAssistant?: string;
+  pendingSearchSources?: SearchSource[];
+  items: Item[];
+};
+
+export function historySearchCards(
+  searches: { id?: string; query?: string; results?: { title?: string; url?: string }[] }[] | undefined,
+): Extract<Item, { kind: "tool" }>[] {
+  const cards: Extract<Item, { kind: "tool" }>[] = [];
+  for (const search of searches ?? []) {
+    if (!search.id) continue;
+    const lines = (search.results ?? []).flatMap((hit) => [hit.title, hit.url].filter(Boolean));
+    cards.push({
+      kind: "tool",
+      id: search.id,
+      name: "web_search",
+      args: search.query ? JSON.stringify({ query: search.query }) : "",
+      readOnly: true,
+      status: "done",
+      output: lines.join("\n"),
+    });
+  }
+  return cards;
+}
+
+export function historySearchSources(
+  searches: { results?: { title?: string; url?: string }[] }[] | undefined,
+): SearchSource[] | undefined {
+  const sources = searchSourcesFromHistory(searches);
+  return sources.length > 0 ? sources : undefined;
+}
+
+export function historySearchAndAnswer(
+  id: string,
+  m: { content: string; reasoning?: string; workDurationMs?: number; memoryCitations?: Extract<Item, { kind: "assistant" }>["memoryCitations"]; serverSearch?: { id?: string; query?: string; results?: { title?: string; url?: string }[] }[] },
+): Item[] {
+  const out: Item[] = historySearchCards(m.serverSearch);
+  const searchSources = historySearchSources(m.serverSearch);
+  if (m.content.trim() !== "" || (m.reasoning ?? "").trim() !== "" || searchSources) {
+    out.push({
+      kind: "assistant",
+      id,
+      text: m.content,
+      reasoning: m.reasoning ?? "",
+      streaming: false,
+      workDurationMs: m.workDurationMs,
+      memoryCitations: m.memoryCitations,
+      searchSources,
+    });
+  }
+  return out;
+}
+
+export function attachSearchSources<T extends SearchState>(s: T, sources: SearchSource[]): T {
+  if (sources.length === 0) return s;
+  const pendingSearchSources = mergeSearchSources(s.pendingSearchSources, sources);
+  if (!s.currentAssistant) return { ...s, pendingSearchSources };
+  return {
+    ...s,
+    pendingSearchSources,
+    items: s.items.map((it) =>
+      it.kind === "assistant" && it.id === s.currentAssistant
+        ? { ...it, searchSources: mergeSearchSources(it.searchSources, sources) }
+        : it,
+    ),
+  };
+}
+
+export function isBatchedReadOnlyTool(name: string, readOnly: boolean): boolean {
+  return readOnly && name !== "todo_write" && name !== "web_search";
+}
+
+export function attachWebSearchOutput<T extends SearchState>(s: T, name: string, output?: string, err?: string): T {
+  if (name !== "web_search" || !output || err) return s;
+  return attachSearchSources(s, parseSearchSources(output));
+}
