+++ desktop/frontend/src/lib/types.ts
@@ -1,5 +1,6 @@
 // Wire contract — mirrors desktop/wire.go (itself mirroring internal/serve/wire.go).
 // One event channel carries every kind; `kind` discriminates the payload.
+import type { HistoryServerSearch } from "./searchSources";
 import type { Todo } from "./tools";
 import type { ContextMaintenanceInfo, WireContextMaintenance } from "./contextMaintenanceTypes";
 export type { ContextMaintenanceInfo, ContextMaintenanceReceipt, WireContextMaintenance } from "./contextMaintenanceTypes";
@@ -698,12 +699,6 @@ export interface HistoryMessage {
   archive?: string;
   decisionReceipt?: WireDecisionReceipt;
   serverSearch?: HistoryServerSearch[];
-}
-
-export interface HistoryServerSearch {
-  id: string;
-  query?: string;
-  results?: { title?: string; url?: string }[];
 }
 
 export interface HistoryToolCall {
