+++ internal/agent/agent.go
@@ -1791,8 +1791,7 @@ func (a *Agent) streamWithFrozen(ctx context.Context, turn int, sink event.Sink,
 	var reasoningID, reasoningStatus string // Responses reasoning item id/status (meta chunk)
 	var calls []provider.ToolCall
 	var responsesItems []json.RawMessage
-	var serverSearch []provider.ServerSearchCall
-	searchDispatched := map[string]bool{}
+	search := newSearchTurn()
 	var partialCalls []provider.ToolCall
 	var usage *provider.Usage
 	var partialToolStarted bool
@@ -1804,7 +1803,7 @@ func (a *Agent) streamWithFrozen(ctx context.Context, turn int, sink event.Sink,
 		return streamedTurn{
 			text: text.String(), reasoning: stored, signature: signature,
 			reasoningID: reasoningID, reasoningStatus: reasoningStatus,
-			calls: calls, responsesItems: responsesItems, serverSearch: serverSearch, usage: usage,
+			calls: calls, responsesItems: responsesItems, serverSearch: search.calls, usage: usage,
 			partialToolStarted: partialToolStarted, partialCalls: partialCalls,
 			maxArgChars: maxArgChars, err: err,
 		}
@@ -1876,7 +1875,7 @@ func (a *Agent) streamWithFrozen(ctx context.Context, turn int, sink event.Sink,
 				return streamedTurn{
 					text: finalText, reasoning: finalReasoning, signature: signature,
 					reasoningID: reasoningID, reasoningStatus: reasoningStatus,
-					calls: calls, responsesItems: responsesItems, serverSearch: serverSearch, usage: usage,
+					calls: calls, responsesItems: responsesItems, serverSearch: search.calls, usage: usage,
 					partialCalls: partialCalls, maxArgChars: maxArgChars,
 				}
 			}
@@ -1951,25 +1950,7 @@ func (a *Agent) streamWithFrozen(ctx context.Context, turn int, sink event.Sink,
 				responsesItems = append(responsesItems, append(json.RawMessage(nil), chunk.ResponsesItem...))
 			}
 		case provider.ChunkServerSearch:
-			if chunk.ServerSearch == nil {
-				break
-			}
-			serverSearch = provider.MergeServerSearch(serverSearch, *chunk.ServerSearch)
-			call := chunk.ServerSearch
-			args := provider.FormatServerSearchArgs(call.Query)
-			completed := len(call.Results) > 0 || len(call.Raw) > 0
-			if !searchDispatched[call.ID] || !completed {
-				sink.Emit(event.Event{Kind: event.ToolDispatch, Tool: event.Tool{
-					ID: call.ID, Name: "web_search", Args: args, ReadOnly: true, AttemptID: attemptID,
-				}})
-				searchDispatched[call.ID] = true
-			}
-			if !completed {
-				break
-			}
-			sink.Emit(event.Event{Kind: event.ToolResult, Tool: event.Tool{
-				ID: call.ID, Name: "web_search", Args: args, Output: provider.FormatServerSearchOutput(call.Results), ReadOnly: true, AttemptID: attemptID,
-			}})
+			search.onChunk(sink, chunk, attemptID)
 		case provider.ChunkUsage:
 			usage, a.turn.lastReasoning = chunk.Usage, chunk.Usage.ReasoningTokens
 			a.storeLatestRequestUsage(chunk.Usage)
