+++ internal/cli/chat_tui.go
@@ -116,10 +116,8 @@ type chatTUI struct {
 	// todoArgs is the latest todo_write call's raw args; it drives the task list
 	// pinned just above the input (see renderTodoPanel). "" when there's no list.
 	// Persists across turns until the work completes or a new session starts.
-	todoArgs string
-	// searchSources is the latest provider web_search hit list, printed as
-	// post-answer footnotes and cleared when the turn settles.
-	searchSources []provider.ServerSearchHit
+	todoArgs      string
+	searchSources []provider.ServerSearchHit // post-answer footnotes; cleared when the turn settles
 
 	// marker rides in outgoing user messages so the cache-stable prompt prefix is
 	// left untouched.
@@ -4500,10 +4498,7 @@ func (m *chatTUI) ingestEvent(e event.Event) {
 			m.pending.Reset()
 			m.pending.WriteString(e.Text)
 		}
-		if notes := provider.FormatServerSearchFootnotes(m.searchSources); notes != "" {
-			m.pending.WriteString(notes)
-			m.searchSources = nil
-		}
+		m.writeSearchFootnotes()
 		m.commitReasoning()
 		m.commitPending()
 
@@ -4558,9 +4553,7 @@ func (m *chatTUI) ingestEvent(e event.Event) {
 		if e.Tool.Name == "todo_write" && e.Tool.Err == "" {
 			m.todoArgs = e.Tool.Args
 		}
-		if e.Tool.Name == "web_search" && e.Tool.Err == "" {
-			m.searchSources = provider.ParseServerSearchOutput(e.Tool.Output)
-		}
+		m.rememberSearchResult(e.Tool)
 		if e.Tool.Err != "" {
 			m.finalizeStreamed()
 			label := shellToolDisplayName(e.Tool.Name, e.Tool.Execution)
@@ -4707,10 +4700,7 @@ func (m *chatTUI) ingestEvent(e event.Event) {
 		// and gate a plan-mode proposal on the user's approval. Autosave already
 		// happened in Controller so every frontend shares the same activity-time
 		// semantics.
-		if notes := provider.FormatServerSearchFootnotes(m.searchSources); notes != "" {
-			m.pending.WriteString(notes)
-			m.searchSources = nil
-		}
+		m.writeSearchFootnotes()
 		m.commitReasoning()
 		m.commitPending()
 		// The bubble was echoed on Enter and an un-sent turn is swallowed above
