+++ internal/provider/provider.go
@@ -73,15 +73,10 @@ type Message struct {
 	// Round-tripped alongside ReasoningContent.
 	ReasoningSignature string     `json:"reasoning_signature,omitempty"`
 	ToolCalls          []ToolCall `json:"tool_calls,omitempty"` // set by assistant
-	// ResponsesItems preserves provider-issued Responses API output items that
-	// must be replayed on a stateless follow-up. Today only DeepSeek
-	// web_search_call items use this path; other providers ignore the field.
-	// Keeping the opaque JSON on the assistant turn makes resume/restart safe,
-	// while omitempty keeps old session files byte-compatible when unused.
-	ResponsesItems []json.RawMessage `json:"responses_items,omitempty"`
-	// ServerSearch holds provider-executed web_search turns for cards and
-	// Anthropic block replay. omitempty keeps old session files unchanged.
-	ServerSearch    []ServerSearchCall `json:"server_search,omitempty"`
+	// ResponsesItems preserves provider-issued Responses API output items for
+	// stateless replay. omitempty keeps old session files byte-compatible.
+	ResponsesItems  []json.RawMessage  `json:"responses_items,omitempty"`
+	ServerSearch    []ServerSearchCall `json:"server_search,omitempty"`   // cards + Anthropic replay; omitempty
 	ToolCallID      string             `json:"tool_call_id,omitempty"`    // links a tool result to its call
 	Name            string             `json:"name,omitempty"`            // tool message: tool name
 	MemoryCitations []MemoryCitation   `json:"memoryCitations,omitempty"` // local UI metadata; provider requests ignore it
