+++ internal/provider/responses/responses.go
@@ -650,13 +650,8 @@ func (c *client) readStream(ctx context.Context, resp *http.Response, out chan<-
 						seenSearchItems[key] = struct{}{}
 						raw := append(json.RawMessage(nil), event.Item.Raw...)
 						responsesItems = append(responsesItems, raw)
-						if !sendChunk(ctx, out, provider.Chunk{Type: provider.ChunkResponsesItem, ResponsesItem: raw}) {
+						if !emitSearchReplay(ctx, out, raw) {
 							return
-						}
-						if search := provider.ServerSearchFromResponsesItem(raw); search != nil {
-							if !sendChunk(ctx, out, provider.Chunk{Type: provider.ChunkServerSearch, ServerSearch: search}) {
-								return
-							}
 						}
 					}
 				}
